Added palettes
I've added a new palette definition file and he means to rotate them. Looking at the various palettes, some look better suited to sunrises and some look better suited for sunsets so I broke them apart. Also, added segments to the strip so the middle section is "ahead" of the sides to create a hotspot. Not sure how that will work.