Updated palettes to finish by 240 instead of 255
For whatever reason, anything in the palette above 240 appears as full white. I suspect it's the fastled fork I'm using to get support for the RGBW strip. Easy fix is to stop the animation at 240 and make sure the palettes all reach their "destination" of 0,0,0 by 240 in the palette index instead of 255.